+++ Cargo.toml
@@ -6,10 +6,10 @@ edition = "2021"
 [workspace]
 members = [
     ".",
-    "libs/whip-whep",
     "libs/cli",
-    "tools/rtp2whip",
-    "tools/whep2rtp",
+    "libs/libwish",
+    "tools/whepfrom",
+    "tools/whipinto",
 ]
 
 [dependencies]
