+++ README.md
@@ -85,10 +85,18 @@ For Example:
 ffmpeg -> rtp2whip -> live777 -> whep2rtp -> ffplay
 ```
 
-### rtp2whip
+### whipinto
+
+This tool is `rtp2whip`
+
+Build:
 
 ```bash
-cargo run --package=rtp2whip -- -c vp8 -u http://localhost:3000/whip/777 --port 5003
+cargo build --package=whipinto --release
+```
+
+```bash
+whipinto -c vp8 -u http://localhost:3000/whip/777 --port 5003
 ```
 
 Video Test Src
@@ -100,14 +108,22 @@ ffmpeg -re -f lavfi -i testsrc=size=640x480:rate=30 -vcodec libvpx -f rtp 'rtp:/
 So. We support parameter `command`, You can use this:
 
 ```bash
-cargo run --package=rtp2whip -- -c vp8 -u http://localhost:3000/whip/777 --command \
+cargo run --package=whipinto -- -c vp8 -u http://localhost:3000/whip/777 --command \
 "ffmpeg -re -f lavfi -i testsrc=size=640x480:rate=30 -vcodec libvpx -cpu-used 5 -deadline 1 -g 10 -error-resilient 1 -auto-alt-ref 1 -f rtp 'rtp://127.0.0.1:{port}?pkt_size=1200'"
 ```
 
-### whep2rtp
+### whepfrom
+
+Build:
 
 ```bash
-cargo run --package=whep2rtp -- -c vp8 -u http://localhost:3000/whep/777 -t localhost:5004
+cargo build --package=whepfrom --release
+```
+
+Use WHEP protocol pull stream convert rtp sender
+
+```bash
+whepfrom -c vp8 -u http://localhost:3000/whep/777 -t localhost:5004
 ```
 
 For ffplay. You Need a sdp file
@@ -130,7 +146,7 @@ ffplay -protocol_whitelist rtp,file,udp -i stream.sdp
 So. You can use this:
 
 ```bash
-cargo run --package=whep2rtp -- -c vp8 -u http://localhost:3000/whep/777 -t 127.0.0.1:5004 --command 'ffplay -protocol_whitelist rtp,file,udp -i stream.sdp'
+cargo run --package=whepfrom -- -c vp8 -u http://localhost:3000/whep/777 -t 127.0.0.1:5004 --command 'ffplay -protocol_whitelist rtp,file,udp -i stream.sdp'
 ```
 
 ## Sponsors
