+++ libs/libwish/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
-name = "whep2rtp"
+name = "libwish"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+crate-type = ["lib"]
 
 [dependencies]
-clap = { version = "4.4.1", features = ["derive"] }
 webrtc = { git = "https://github.com/webrtc-rs/webrtc" }
 anyhow = "1.0"
-tokio = { version = "1.30", features = ["full"] }
-cli = { path = "../../libs/cli" }
-whip-whep = { path = "../../libs/whip-whep" }
+reqwest = "0.11.20"
+parse_link_header = "0.3"
+base64 = "0.21.3"
