+++ tools/whepfrom/Cargo.toml
@@ -1,15 +1,14 @@
 [package]
-name = "whip-whep"
+name = "whepfrom"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-crate-type = ["lib"]
 
 [dependencies]
+clap = { version = "4.4.1", features = ["derive"] }
 webrtc = { git = "https://github.com/webrtc-rs/webrtc" }
 anyhow = "1.0"
-reqwest = "0.11.20"
-parse_link_header = "0.3"
-base64 = "0.21.3"
+tokio = { version = "1.30", features = ["full"] }
+cli = { path = "../../libs/cli" }
+libwish = { path = "../../libs/libwish" }
